+++ just_audio_background/lib/just_audio_background.dart
@@ -6,6 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:just_audio_platform_interface/just_audio_platform_interface.dart';
 import 'package:rxdart/rxdart.dart';
+import 'package:synchronized/synchronized.dart';
 
 export 'package:audio_service/audio_service.dart' show MediaItem;
 
@@ -121,20 +122,26 @@ class _JustAudioBackgroundPlugin extends JustAudioPlatform {
 
   @override
   Future<AudioPlayerPlatform> init(InitRequest request) async {
-    if (_player != null) {
-      throw PlatformException(code: "error", message: "just_audio_background supports only a single player instance");
+    if (_playerId != null) {
+      throw PlatformException(
+        code: "error",
+        message: "just_audio_background supports only a single player instance",
+      );
     }
-    _player = _JustAudioPlayer(
-      initRequest: request,
-    );
+    _playerId = request.id;
+    _player ??= _JustAudioPlayer(initRequest: request);
     return _player!;
   }
 
   @override
-  Future<DisposePlayerResponse> disposePlayer(DisposePlayerRequest request) async {
-    final player = _player;
-    _player = null;
-    await player?.release();
+  Future<DisposePlayerResponse> disposePlayer(
+      DisposePlayerRequest request) async {
+    if (request.id == _playerId) {
+      _playerId = null;
+      final player = _player;
+      _player = null;
+      await player?.release();
+    }
     return DisposePlayerResponse();
   }
 
@@ -147,82 +154,48 @@ class _JustAudioBackgroundPlugin extends JustAudioPlatform {
   }
 }
 
+final _PlayerAudioHandler _playerAudioHandler = _PlayerAudioHandler();
+
 class _JustAudioPlayer extends AudioPlayerPlatform {
   final InitRequest initRequest;
-  final eventController = StreamController<PlaybackEventMessage>.broadcast();
-  final playerDataController = StreamController<PlayerDataMessage>.broadcast();
-  bool? _playing;
-  IcyMetadataMessage? _icyMetadata;
-  int? _androidAudioSessionId;
-  late final _PlayerAudioHandler _playerAudioHandler;
+  final eventController =
+      StreamController<PlaybackEventMessage>.broadcast(sync: true);
+  final playerDataController =
+      StreamController<PlayerDataMessage>.broadcast(sync: true);
 
   _JustAudioPlayer({required this.initRequest}) : super(initRequest.id) {
-    _playerAudioHandler = _PlayerAudioHandler(initRequest);
+    eventController.onCancel = _playerAudioHandler.cancelStreamSubscriptions;
+    _playerAudioHandler._initPlayer(initRequest);
     _audioHandler.inner = _playerAudioHandler;
-    _audioHandler.playbackState.listen((playbackState) {
-      broadcastPlaybackEvent();
-    });
-    _audioHandler.customEvent.listen((dynamic event) {
-      switch (event['type']) {
-        case 'icyMetadata':
-          _icyMetadata = event['value'] as IcyMetadataMessage?;
-          broadcastPlaybackEvent();
-          break;
-        case 'androidAudioSessionId':
-          _androidAudioSessionId = event['value'] as int?;
-          broadcastPlaybackEvent();
-          break;
-      }
-    });
-    _audioHandler.mediaItem.listen((mediaItem) {
-      if (mediaItem == null) return;
-      broadcastPlaybackEvent();
+    _audioHandler.customEvent
+        .whereType<PlaybackEventMessage>()
+        .listen(eventController.add);
+    _audioHandler.customEvent
+        .whereType<_PlayingEvent>()
+        .map((event) => event.playing)
+        .distinct()
+        .listen((playing) {
+      playerDataController.add(PlayerDataMessage(playing: playing));
     });
   }
 
   PlaybackState get playbackState => _audioHandler.playbackState.nvalue!;
 
   Future<void> release() async {
-    eventController.close();
     await _audioHandler.stop();
   }
 
-  void broadcastPlaybackEvent() {
-    if (eventController.isClosed) return;
-    eventController.add(PlaybackEventMessage(
-      //processingState: playbackState.processingState,
-      processingState: {
-        AudioProcessingState.idle: ProcessingStateMessage.idle,
-        AudioProcessingState.loading: ProcessingStateMessage.loading,
-        AudioProcessingState.ready: ProcessingStateMessage.ready,
-        AudioProcessingState.buffering: ProcessingStateMessage.buffering,
-        AudioProcessingState.completed: ProcessingStateMessage.completed,
-        AudioProcessingState.error: ProcessingStateMessage.idle,
-      }[playbackState.processingState]!,
-      updatePosition: playbackState.position,
-      updateTime: playbackState.updateTime,
-      bufferedPosition: playbackState.bufferedPosition,
-      icyMetadata: _icyMetadata,
-      duration: _playerAudioHandler.currentMediaItem?.duration,
-      currentIndex: playbackState.queueIndex,
-      androidAudioSessionId: _androidAudioSessionId,
-    ));
-    if (playbackState.playing != _playing) {
-      _playing = playbackState.playing;
-      playerDataController.add(PlayerDataMessage(
-        playing: playbackState.playing,
-      ));
-    }
-  }
+  @override
+  Stream<PlaybackEventMessage> get playbackEventMessageStream =>
+      eventController.stream;
 
   @override
-  Stream<PlaybackEventMessage> get playbackEventMessageStream => eventController.stream;
+  Stream<PlayerDataMessage> get playerDataMessageStream =>
+      playerDataController.stream;
 
   @override
-  Stream<PlayerDataMessage> get playerDataMessageStream => playerDataController.stream;
-
-  @override
-  Future<LoadResponse> load(LoadRequest request) => _playerAudioHandler.customLoad(request);
+  Future<LoadResponse> load(LoadRequest request) =>
+      _playerAudioHandler.customLoad(request);
 
   @override
   Future<PlayResponse> play(PlayRequest request) async {
@@ -237,7 +210,8 @@ class _JustAudioPlayer extends AudioPlayerPlatform {
   }
 
   @override
-  Future<SetVolumeResponse> setVolume(SetVolumeRequest request) => _playerAudioHandler.customSetVolume(request);
+  Future<SetVolumeResponse> setVolume(SetVolumeRequest request) =>
+      _playerAudioHandler.customSetVolume(request);
 
   @override
   Future<SetSpeedResponse> setSpeed(SetSpeedRequest request) async {
@@ -252,59 +226,79 @@ class _JustAudioPlayer extends AudioPlayerPlatform {
   }
 
   @override
-  Future<SetSkipSilenceResponse> setSkipSilence(SetSkipSilenceRequest request) async {
+  Future<SetSkipSilenceResponse> setSkipSilence(
+      SetSkipSilenceRequest request) async {
     await _playerAudioHandler.customSetSkipSilence(request);
     return SetSkipSilenceResponse();
   }
 
   @override
   Future<SetLoopModeResponse> setLoopMode(SetLoopModeRequest request) async {
-    await _audioHandler.setRepeatMode(AudioServiceRepeatMode.values[request.loopMode.index]);
+    await _audioHandler
+        .setRepeatMode(AudioServiceRepeatMode.values[request.loopMode.index]);
     return SetLoopModeResponse();
   }
 
   @override
-  Future<SetShuffleModeResponse> setShuffleMode(SetShuffleModeRequest request) async {
-    await _audioHandler.setShuffleMode(AudioServiceShuffleMode.values[request.shuffleMode.index]);
+  Future<SetShuffleModeResponse> setShuffleMode(
+      SetShuffleModeRequest request) async {
+    await _audioHandler.setShuffleMode(
+        AudioServiceShuffleMode.values[request.shuffleMode.index]);
     return SetShuffleModeResponse();
   }
 
   @override
-  Future<SetShuffleOrderResponse> setShuffleOrder(SetShuffleOrderRequest request) =>
+  Future<SetShuffleOrderResponse> setShuffleOrder(
+          SetShuffleOrderRequest request) =>
       _playerAudioHandler.customSetShuffleOrder(request);
 
   @override
-  Future<SetWebCrossOriginResponse> setWebCrossOrigin(SetWebCrossOriginRequest request) async {
+  Future<SetWebCrossOriginResponse> setWebCrossOrigin(
+      SetWebCrossOriginRequest request) async {
     _playerAudioHandler.customSetWebCrossOrigin(request);
     return SetWebCrossOriginResponse();
   }
 
   @override
-  Future<SeekResponse> seek(SeekRequest request) => _playerAudioHandler.customPlayerSeek(request);
+  Future<SetWebSinkIdResponse> setWebSinkId(SetWebSinkIdRequest request) {
+    _playerAudioHandler.customSetWebSinkId(request);
+    throw SetWebSinkIdResponse();
+  }
 
   @override
-  Future<ConcatenatingInsertAllResponse> concatenatingInsertAll(ConcatenatingInsertAllRequest request) =>
+  Future<SeekResponse> seek(SeekRequest request) =>
+      _playerAudioHandler.customPlayerSeek(request);
+
+  @override
+  Future<ConcatenatingInsertAllResponse> concatenatingInsertAll(
+          ConcatenatingInsertAllRequest request) =>
       _playerAudioHandler.customConcatenatingInsertAll(request);
 
   @override
-  Future<ConcatenatingRemoveRangeResponse> concatenatingRemoveRange(ConcatenatingRemoveRangeRequest request) =>
+  Future<ConcatenatingRemoveRangeResponse> concatenatingRemoveRange(
+          ConcatenatingRemoveRangeRequest request) =>
       _playerAudioHandler.customConcatenatingRemoveRange(request);
 
   @override
-  Future<ConcatenatingMoveResponse> concatenatingMove(ConcatenatingMoveRequest request) =>
+  Future<ConcatenatingMoveResponse> concatenatingMove(
+          ConcatenatingMoveRequest request) =>
       _playerAudioHandler.customConcatenatingMove(request);
 
   @override
-  Future<SetAndroidAudioAttributesResponse> setAndroidAudioAttributes(SetAndroidAudioAttributesRequest request) =>
+  Future<SetAndroidAudioAttributesResponse> setAndroidAudioAttributes(
+          SetAndroidAudioAttributesRequest request) =>
       _playerAudioHandler.customSetAndroidAudioAttributes(request);
 
   @override
-  Future<SetAutomaticallyWaitsToMinimizeStallingResponse> setAutomaticallyWaitsToMinimizeStalling(
-          SetAutomaticallyWaitsToMinimizeStallingRequest request) =>
-      _playerAudioHandler.customSetAutomaticallyWaitsToMinimizeStalling(request);
+  Future<SetAutomaticallyWaitsToMinimizeStallingResponse>
+      setAutomaticallyWaitsToMinimizeStalling(
+              SetAutomaticallyWaitsToMinimizeStallingRequest request) =>
+          _playerAudioHandler
+              .customSetAutomaticallyWaitsToMinimizeStalling(request);
 
   @override
-  Future<AndroidEqualizerBandSetGainResponse> androidEqualizerBandSetGain(AndroidEqualizerBandSetGainRequest request) =>
+  Future<AndroidEqualizerBandSetGainResponse> androidEqualizerBandSetGain(
+          AndroidEqualizerBandSetGainRequest request) =>
       _playerAudioHandler.customAndroidEqualizerBandSetGain(request);
 
   @override
@@ -313,31 +307,41 @@ class _JustAudioPlayer extends AudioPlayerPlatform {
       _playerAudioHandler.customAndroidEqualizerGetParameters(request);
 
   @override
-  Future<AndroidLoudnessEnhancerSetTargetGainResponse> androidLoudnessEnhancerSetTargetGain(
-          AndroidLoudnessEnhancerSetTargetGainRequest request) =>
-      _playerAudioHandler.customAndroidLoudnessEnhancerSetTargetGain(request);
+  Future<AndroidLoudnessEnhancerSetTargetGainResponse>
+      androidLoudnessEnhancerSetTargetGain(
+              AndroidLoudnessEnhancerSetTargetGainRequest request) =>
+          _playerAudioHandler
+              .customAndroidLoudnessEnhancerSetTargetGain(request);
 
   @override
-  Future<AudioEffectSetEnabledResponse> audioEffectSetEnabled(AudioEffectSetEnabledRequest request) =>
+  Future<AudioEffectSetEnabledResponse> audioEffectSetEnabled(
+          AudioEffectSetEnabledRequest request) =>
       _playerAudioHandler.customAudioEffectSetEnabled(request);
 
   @override
-  Future<SetAllowsExternalPlaybackResponse> setAllowsExternalPlayback(SetAllowsExternalPlaybackRequest request) =>
+  Future<SetAllowsExternalPlaybackResponse> setAllowsExternalPlayback(
+          SetAllowsExternalPlaybackRequest request) =>
       _playerAudioHandler.customSetAllowsExternalPlayback(request);
 
   @override
   Future<SetCanUseNetworkResourcesForLiveStreamingWhilePausedResponse>
       setCanUseNetworkResourcesForLiveStreamingWhilePaused(
-              SetCanUseNetworkResourcesForLiveStreamingWhilePausedRequest request) =>
-          _playerAudioHandler.customSetCanUseNetworkResourcesForLiveStreamingWhilePaused(request);
+              SetCanUseNetworkResourcesForLiveStreamingWhilePausedRequest
+                  request) =>
+          _playerAudioHandler
+              .customSetCanUseNetworkResourcesForLiveStreamingWhilePaused(
+                  request);
 
   @override
-  Future<SetPreferredPeakBitRateResponse> setPreferredPeakBitRate(SetPreferredPeakBitRateRequest request) =>
+  Future<SetPreferredPeakBitRateResponse> setPreferredPeakBitRate(
+          SetPreferredPeakBitRateRequest request) =>
       _playerAudioHandler.customSetPreferredPeakBitRate(request);
 }
 
-class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandler {
-  final _playerCompleter = Completer<AudioPlayerPlatform>();
+class _PlayerAudioHandler extends BaseAudioHandler
+    with QueueHandler, SeekHandler {
+  final _lock = Lock();
+  var _playerCompleter = _ValueCompleter<AudioPlayerPlatform>();
   PlaybackEventMessage _justAudioEvent = PlaybackEventMessage(
     processingState: ProcessingStateMessage.idle,
     updateTime: DateTime.now(),
@@ -360,70 +364,72 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
   List<int> _effectiveIndicesInv = [];
 
   Future<AudioPlayerPlatform> get _player => _playerCompleter.future;
-  int? get index => _justAudioEvent.currentIndex;
+  int? index;
   MediaItem? get currentMediaItem =>
-      index != null && currentQueue != null && index! >= 0 && index! < currentQueue!.length
-          ? currentQueue![index!]
+      index != null && index! >= 0 && index! < currentQueue.length
+          ? currentQueue[index!]
           : null;
 
-  List<MediaItem>? get currentQueue => queue.nvalue;
+  List<MediaItem> get currentQueue => queue.value;
+  StreamSubscription<TrackInfo>? _trackInfoSubscription;
 
-  _PlayerAudioHandler(InitRequest initRequest) {
-    _init(initRequest);
-  }
+  Future<void> _initPlayer(InitRequest initRequest) =>
+      _lock.synchronized(() async {
+        final player = await _platform.init(initRequest);
+        _playerCompleter.complete(player);
+        final playbackEventMessageStream = player.playbackEventMessageStream;
+        _trackInfoSubscription = playbackEventMessageStream
+            .map((event) {
+              index = event.currentIndex ?? _justAudioEvent.currentIndex;
+              _justAudioEvent = event;
+              customEvent.add(event);
+              _broadcastState();
+              return event;
+            })
+            .map((event) => TrackInfo(event.currentIndex, event.duration))
+            .distinct()
+            .debounceTime(const Duration(milliseconds: 100))
+            .map((track) {
+              // Platform may send us a null duration on dispose, which we should
+              // ignore.
+              final currentMediaItem = this.currentMediaItem;
+              if (currentMediaItem != null) {
+                if (track.duration == null &&
+                    currentMediaItem.duration != null) {
+                  return TrackInfo(track.index, currentMediaItem.duration);
+                }
+              }
+              return track;
+            })
+            .distinct()
+            .listen((track) {
+              if (currentMediaItem != null && index != null) {
+                if (track.duration != currentMediaItem!.duration &&
+                    (index! < queue.nvalue!.length && track.duration != null)) {
+                  currentQueue[index!] =
+                      currentQueue[index!].copyWith(duration: track.duration);
+                  queue.add(currentQueue);
+                }
+                mediaItem.add(currentMediaItem!);
+              }
+            }, onError: (Object e, [StackTrace? st]) {});
+      });
 
-  Future<void> _init(InitRequest initRequest) async {
-    final player = await _platform.init(initRequest);
-    _playerCompleter.complete(player);
-    final playbackEventMessageStream = player.playbackEventMessageStream;
-    playbackEventMessageStream.listen((event) {
-      _justAudioEvent = event;
-      _broadcastState();
-    });
-    playbackEventMessageStream.map((event) => event.icyMetadata).distinct().listen((icyMetadata) {
-      customEvent.add({
-        'type': 'icyMetadata',
-        'value': icyMetadata,
-      });
-    });
-    playbackEventMessageStream.map((event) => event.androidAudioSessionId).distinct().listen((audioSessionId) {
-      customEvent.add({
-        'type': 'androidAudioSessionId',
-        'value': audioSessionId,
-      });
-    });
-    playbackEventMessageStream
-        .map((event) => TrackInfo(event.currentIndex, event.duration))
-        .distinct()
-        .debounceTime(const Duration(milliseconds: 100))
-        .map((track) {
-          // Platform may send us a null duration on dispose, which we should
-          // ignore.
-          final currentMediaItem = this.currentMediaItem;
-          if (currentMediaItem != null) {
-            if (track.duration == null && currentMediaItem.duration != null) {
-              return TrackInfo(track.index, currentMediaItem.duration);
-            }
-          }
-          return track;
-        })
-        .distinct()
-        .listen((track) {
-          if (currentMediaItem != null) {
-            if (track.duration != currentMediaItem!.duration &&
-                (index! < queue.nvalue!.length && track.duration != null)) {
-              currentQueue![index!] = currentQueue![index!].copyWith(duration: track.duration);
-              queue.add(currentQueue!);
-            }
-            mediaItem.add(currentMediaItem!);
-          }
-        });
+  Future<void> cancelStreamSubscriptions() async {
+    final trackInfoSubscription = _trackInfoSubscription;
+    if (trackInfoSubscription != null) {
+      _trackInfoSubscription = null;
+      await trackInfoSubscription.cancel();
+    }
   }
 
   @override
   Future<void> updateQueue(List<MediaItem> queue) async {
     this.queue.add(queue);
-    if (mediaItem.nvalue == null && index != null && index! >= 0 && index! < queue.length) {
+    if (mediaItem.nvalue == null &&
+        index != null &&
+        index! >= 0 &&
+        index! < queue.length) {
       mediaItem.add(queue[index!]);
     }
   }
@@ -440,18 +446,24 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
     return LoadResponse(duration: response.duration);
   }
 
-  Future<SetVolumeResponse> customSetVolume(SetVolumeRequest request) async => await (await _player).setVolume(request);
+  Future<SetVolumeResponse> customSetVolume(SetVolumeRequest request) async =>
+      await (await _player).setVolume(request);
 
-  Future<SetSpeedResponse> customSetSpeed(SetSpeedRequest request) async => await (await _player).setSpeed(request);
+  Future<SetSpeedResponse> customSetSpeed(SetSpeedRequest request) async =>
+      await (await _player).setSpeed(request);
 
-  Future<SetPitchResponse> customSetPitch(SetPitchRequest request) async => await (await _player).setPitch(request);
+  Future<SetPitchResponse> customSetPitch(SetPitchRequest request) async =>
+      await (await _player).setPitch(request);
 
-  Future<SetSkipSilenceResponse> customSetSkipSilence(SetSkipSilenceRequest request) async =>
+  Future<SetSkipSilenceResponse> customSetSkipSilence(
+          SetSkipSilenceRequest request) async =>
       await (await _player).setSkipSilence(request);
 
-  Future<SeekResponse> customPlayerSeek(SeekRequest request) async => await (await _player).seek(request);
+  Future<SeekResponse> customPlayerSeek(SeekRequest request) async =>
+      await (await _player).seek(request);
 
-  Future<SetShuffleOrderResponse> customSetShuffleOrder(SetShuffleOrderRequest request) async {
+  Future<SetShuffleOrderResponse> customSetShuffleOrder(
+      SetShuffleOrderRequest request) async {
     _source = request.audioSourceMessage;
     _updateShuffleIndices();
     _broadcastStateIfActive();
@@ -460,13 +472,22 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
     ));
   }
 
-  Future<SetWebCrossOriginResponse> customSetWebCrossOrigin(SetWebCrossOriginRequest request) async {
+  Future<SetWebCrossOriginResponse> customSetWebCrossOrigin(
+      SetWebCrossOriginRequest request) async {
     return await (await _player).setWebCrossOrigin(request);
   }
 
-  Future<ConcatenatingInsertAllResponse> customConcatenatingInsertAll(ConcatenatingInsertAllRequest request) async {
+  Future<SetWebSinkIdResponse> customSetWebSinkId(
+      SetWebSinkIdRequest request) async {
+    return await (await _player).setWebSinkId(request);
+  }
+
+  Future<ConcatenatingInsertAllResponse> customConcatenatingInsertAll(
+      ConcatenatingInsertAllRequest request) async {
     final cat = _source!.findCat(request.id)!;
     cat.children.insertAll(request.index, request.children);
+    cat.shuffleOrder
+        .replaceRange(0, cat.shuffleOrder.length, request.shuffleOrder);
     _updateShuffleIndices();
     _broadcastStateIfActive();
     _updateQueue();
@@ -477,15 +498,21 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
       ConcatenatingRemoveRangeRequest request) async {
     final cat = _source!.findCat(request.id)!;
     cat.children.removeRange(request.startIndex, request.endIndex);
+    cat.shuffleOrder
+        .replaceRange(0, cat.shuffleOrder.length, request.shuffleOrder);
     _updateShuffleIndices();
     _broadcastStateIfActive();
     _updateQueue();
     return await (await _player).concatenatingRemoveRange(request);
   }
 
-  Future<ConcatenatingMoveResponse> customConcatenatingMove(ConcatenatingMoveRequest request) async {
+  Future<ConcatenatingMoveResponse> customConcatenatingMove(
+      ConcatenatingMoveRequest request) async {
     final cat = _source!.findCat(request.id)!;
-    cat.children.insert(request.newIndex, cat.children.removeAt(request.currentIndex));
+    cat.children
+        .insert(request.newIndex, cat.children.removeAt(request.currentIndex));
+    cat.shuffleOrder
+        .replaceRange(0, cat.shuffleOrder.length, request.shuffleOrder);
     _updateShuffleIndices();
     _broadcastStateIfActive();
     _updateQueue();
@@ -496,23 +523,28 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
           SetAndroidAudioAttributesRequest request) async =>
       await (await _player).setAndroidAudioAttributes(request);
 
-  Future<SetAutomaticallyWaitsToMinimizeStallingResponse> customSetAutomaticallyWaitsToMinimizeStalling(
-          SetAutomaticallyWaitsToMinimizeStallingRequest request) async =>
-      await (await _player).setAutomaticallyWaitsToMinimizeStalling(request);
+  Future<SetAutomaticallyWaitsToMinimizeStallingResponse>
+      customSetAutomaticallyWaitsToMinimizeStalling(
+              SetAutomaticallyWaitsToMinimizeStallingRequest request) async =>
+          await (await _player)
+              .setAutomaticallyWaitsToMinimizeStalling(request);
 
   Future<AndroidEqualizerBandSetGainResponse> customAndroidEqualizerBandSetGain(
           AndroidEqualizerBandSetGainRequest request) async =>
       await (await _player).androidEqualizerBandSetGain(request);
 
-  Future<AndroidEqualizerGetParametersResponse> customAndroidEqualizerGetParameters(
-          AndroidEqualizerGetParametersRequest request) async =>
-      await (await _player).androidEqualizerGetParameters(request);
+  Future<AndroidEqualizerGetParametersResponse>
+      customAndroidEqualizerGetParameters(
+              AndroidEqualizerGetParametersRequest request) async =>
+          await (await _player).androidEqualizerGetParameters(request);
 
-  Future<AndroidLoudnessEnhancerSetTargetGainResponse> customAndroidLoudnessEnhancerSetTargetGain(
-          AndroidLoudnessEnhancerSetTargetGainRequest request) async =>
-      await (await _player).androidLoudnessEnhancerSetTargetGain(request);
+  Future<AndroidLoudnessEnhancerSetTargetGainResponse>
+      customAndroidLoudnessEnhancerSetTargetGain(
+              AndroidLoudnessEnhancerSetTargetGainRequest request) async =>
+          await (await _player).androidLoudnessEnhancerSetTargetGain(request);
 
-  Future<AudioEffectSetEnabledResponse> customAudioEffectSetEnabled(AudioEffectSetEnabledRequest request) async =>
+  Future<AudioEffectSetEnabledResponse> customAudioEffectSetEnabled(
+          AudioEffectSetEnabledRequest request) async =>
       await (await _player).audioEffectSetEnabled(request);
 
   Future<SetAllowsExternalPlaybackResponse> customSetAllowsExternalPlayback(
@@ -521,28 +553,33 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
 
   Future<SetCanUseNetworkResourcesForLiveStreamingWhilePausedResponse>
       customSetCanUseNetworkResourcesForLiveStreamingWhilePaused(
-              SetCanUseNetworkResourcesForLiveStreamingWhilePausedRequest request) async =>
-          await (await _player).setCanUseNetworkResourcesForLiveStreamingWhilePaused(request);
+              SetCanUseNetworkResourcesForLiveStreamingWhilePausedRequest
+                  request) async =>
+          await (await _player)
+              .setCanUseNetworkResourcesForLiveStreamingWhilePaused(request);
 
-  Future<SetPreferredPeakBitRateResponse> customSetPreferredPeakBitRate(SetPreferredPeakBitRateRequest request) async =>
+  Future<SetPreferredPeakBitRateResponse> customSetPreferredPeakBitRate(
+          SetPreferredPeakBitRateRequest request) async =>
       await (await _player).setPreferredPeakBitRate(request);
 
   void _updateQueue() {
     assert(sequence.every((source) => source.tag is MediaItem),
-        'Error : When using just_audio_background, you should always use a MediaItem as tag when setting an AudioSource. See AudioSource.uri documentation for more information.');
+        'Error : When using just_audio_background, you should always set a MediaItem tag on every AudioSource. See AudioSource.uri documentation for more information.');
     queue.add(sequence.map((source) => source.tag as MediaItem).toList());
   }
 
   void _updateShuffleIndices() {
     _shuffleIndices = _source?.shuffleIndices ?? [];
-    _effectiveIndices =
-        _shuffleMode != AudioServiceShuffleMode.none ? _shuffleIndices : List.generate(sequence.length, (i) => i);
+    _effectiveIndices = _shuffleMode != AudioServiceShuffleMode.none
+        ? _shuffleIndices
+        : List.generate(sequence.length, (i) => i);
     _shuffleIndicesInv = List.filled(_effectiveIndices.length, 0);
     for (var i = 0; i < _effectiveIndices.length; i++) {
       _shuffleIndicesInv[_effectiveIndices[i]] = i;
     }
-    _effectiveIndicesInv =
-        _shuffleMode != AudioServiceShuffleMode.none ? _shuffleIndicesInv : List.generate(sequence.length, (i) => i);
+    _effectiveIndicesInv = _shuffleMode != AudioServiceShuffleMode.none
+        ? _shuffleIndicesInv
+        : List.generate(sequence.length, (i) => i);
   }
 
   List<IndexedAudioSourceMessage> get sequence => _source?.sequence ?? [];
@@ -550,22 +587,23 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
   List<int> get effectiveIndices => _effectiveIndices;
   List<int> get shuffleIndicesInv => _shuffleIndicesInv;
   List<int> get effectiveIndicesInv => _effectiveIndicesInv;
-  int get nextIndex => getRelativeIndex(1);
-  int get previousIndex => getRelativeIndex(-1);
-  bool get hasNext => nextIndex != -1;
-  bool get hasPrevious => previousIndex != -1;
+  int? get nextIndex => getRelativeIndex(1);
+  int? get previousIndex => getRelativeIndex(-1);
+  bool get hasNext => nextIndex != null;
+  bool get hasPrevious => previousIndex != null;
 
-  int getRelativeIndex(int offset) {
-    if (_repeatMode == AudioServiceRepeatMode.one) return index!;
-    if (effectiveIndices.isEmpty) return -1;
-    if (index! >= effectiveIndicesInv.length) return -1;
+  int? getRelativeIndex(int offset) {
+    if (currentQueue.isEmpty || index == null) return null;
+    if (_repeatMode == AudioServiceRepeatMode.one) return index;
+    if (effectiveIndices.isEmpty) return null;
+    if (index! >= effectiveIndicesInv.length) return null;
     final invPos = effectiveIndicesInv[index!];
     var newInvPos = invPos + offset;
     if (newInvPos >= effectiveIndices.length || newInvPos < 0) {
       if (_repeatMode == AudioServiceRepeatMode.all) {
         newInvPos %= effectiveIndices.length;
       } else {
-        return -1;
+        return null;
       }
     }
     final result = effectiveIndices[newInvPos];
@@ -580,29 +618,34 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
   @override
   Future<void> skipToNext() async {
     if (hasNext) {
-      await skipToQueueItem(nextIndex);
+      await skipToQueueItem(nextIndex!);
     }
   }
 
   @override
   Future<void> skipToPrevious() async {
     if (hasPrevious) {
-      await skipToQueueItem(previousIndex);
+      await skipToQueueItem(previousIndex!);
     }
   }
 
   @override
   Future<void> play() async {
-    _updatePosition();
-    _playing = true;
-    _broadcastState();
-    await (await _player).play(PlayRequest());
+    if (_justAudioEvent.processingState == ProcessingStateMessage.completed) {
+      await skipToQueueItem(0);
+    }
+    if (!_playing) {
+      _updatePosition();
+      customEvent.add(_PlayingEvent(_playing = true));
+      _broadcastState();
+      await (await _player).play(PlayRequest());
+    }
   }
 
   @override
   Future<void> pause() async {
     _updatePosition();
-    _playing = false;
+    customEvent.add(_PlayingEvent(_playing = false));
     _broadcastState();
     await (await _player).pause(PauseRequest());
   }
@@ -615,7 +658,8 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
   }
 
   @override
-  Future<void> seek(Duration position) async => await (await _player).seek(SeekRequest(position: position));
+  Future<void> seek(Duration position) async =>
+      await (await _player).seek(SeekRequest(position: position));
 
   @override
   Future<void> setSpeed(double speed) async {
@@ -624,7 +668,8 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
   }
 
   @override
-  Future<void> fastForward() => _seekRelative(AudioService.config.fastForwardInterval);
+  Future<void> fastForward() =>
+      _seekRelative(AudioService.config.fastForwardInterval);
 
   @override
   Future<void> rewind() => _seekRelative(-AudioService.config.rewindInterval);
@@ -639,8 +684,9 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
   Future<void> setRepeatMode(AudioServiceRepeatMode repeatMode) async {
     _repeatMode = repeatMode;
     _broadcastStateIfActive();
-    (await _player).setLoopMode(
-        SetLoopModeRequest(loopMode: LoopModeMessage.values[min(LoopModeMessage.values.length - 1, repeatMode.index)]));
+    (await _player).setLoopMode(SetLoopModeRequest(
+        loopMode: LoopModeMessage
+            .values[min(LoopModeMessage.values.length - 1, repeatMode.index)]));
   }
 
   @override
@@ -649,32 +695,31 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
     _updateShuffleIndices();
     _broadcastStateIfActive();
     (await _player).setShuffleMode(SetShuffleModeRequest(
-        shuffleMode: ShuffleModeMessage.values[min(ShuffleModeMessage.values.length - 1, shuffleMode.index)]));
+        shuffleMode: ShuffleModeMessage.values[
+            min(ShuffleModeMessage.values.length - 1, shuffleMode.index)]));
   }
 
   @override
-  Future<void> stop() async {
-    if (_justAudioEvent.processingState == ProcessingStateMessage.idle) {
-      return;
-    }
-    _updatePosition();
-    _playing = false;
-    _broadcastState();
-    // TODO: We should really stop listening to events here to mimic
-    // just_audio's behaviour. E.g. if stop() was called, we actually want to
-    // keep the state around even though the platform may be disposing its own
-    // state.
-    _platform.disposePlayer(DisposePlayerRequest(id: (await _player).id));
-    _justAudioEvent = _justAudioEvent.copyWith(
-      processingState: ProcessingStateMessage.idle,
-    );
-  }
+  Future<void> stop() => _lock.synchronized(() async {
+        final player = _playerCompleter.value;
+        if (player == null) return;
+        _updatePosition();
+        customEvent.add(_PlayingEvent(_playing = false));
+        _justAudioEvent = _justAudioEvent.copyWith(
+          processingState: ProcessingStateMessage.idle,
+        );
+        _broadcastState();
+        _playerCompleter = _ValueCompleter<AudioPlayerPlatform>();
+        await _platform.disposePlayer(DisposePlayerRequest(id: player.id));
+      });
 
   Duration get currentPosition {
-    if (_playing && _justAudioEvent.processingState == ProcessingStateMessage.ready) {
+    if (_playing &&
+        _justAudioEvent.processingState == ProcessingStateMessage.ready) {
       return Duration(
           milliseconds: (_justAudioEvent.updatePosition.inMilliseconds +
-                  ((DateTime.now().millisecondsSinceEpoch - _justAudioEvent.updateTime.millisecondsSinceEpoch) *
+                  ((DateTime.now().millisecondsSinceEpoch -
+                          _justAudioEvent.updateTime.millisecondsSinceEpoch) *
                       _speed))
               .toInt());
     } else {
@@ -700,9 +745,9 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
   void _seekContinuously(bool begin, int direction) {
     _seeker?.stop();
     if (begin) {
-      _seeker =
-          _Seeker(this, Duration(seconds: 10 * direction), const Duration(seconds: 1), currentMediaItem!.duration!)
-            ..start();
+      _seeker = _Seeker(this, Duration(seconds: 10 * direction),
+          const Duration(seconds: 1), currentMediaItem!.duration!)
+        ..start();
     }
   }
 
@@ -727,20 +772,30 @@ class _PlayerAudioHandler extends BaseAudioHandler with QueueHandler, SeekHandle
         MediaAction.seekForward,
         MediaAction.seekBackward,
       },
-      androidCompactActionIndices:
-          List.generate(controls.length, (i) => i).where((i) => controls[i].action != MediaAction.stop).toList(),
-      processingState: const {
-        ProcessingStateMessage.idle: AudioProcessingState.idle,
-        ProcessingStateMessage.loading: AudioProcessingState.loading,
-        ProcessingStateMessage.buffering: AudioProcessingState.buffering,
-        ProcessingStateMessage.ready: AudioProcessingState.ready,
-        ProcessingStateMessage.completed: AudioProcessingState.completed,
-      }[_justAudioEvent.processingState]!,
-      playing: _playing,
+      androidCompactActionIndices: List.generate(controls.length, (i) => i)
+          .where((i) => controls[i].action != MediaAction.stop)
+          .toList(),
+      processingState: _justAudioEvent.errorCode != null
+          ? AudioProcessingState.error
+          : const {
+                ProcessingStateMessage.idle: AudioProcessingState.idle,
+                ProcessingStateMessage.loading: AudioProcessingState.loading,
+                ProcessingStateMessage.buffering:
+                    AudioProcessingState.buffering,
+                ProcessingStateMessage.ready: AudioProcessingState.ready,
+                ProcessingStateMessage.completed:
+                    AudioProcessingState.completed,
+              }[_justAudioEvent.processingState] ??
+              AudioProcessingState.idle,
+      playing: _playing &&
+          !{ProcessingStateMessage.idle, ProcessingStateMessage.completed}
+              .contains(_justAudioEvent.processingState),
       updatePosition: currentPosition,
       bufferedPosition: _justAudioEvent.bufferedPosition,
       speed: _speed,
       queueIndex: _justAudioEvent.currentIndex,
+      errorCode: _justAudioEvent.errorCode,
+      errorMessage: _justAudioEvent.errorMessage,
     ));
   }
 }
@@ -794,7 +849,8 @@ extension _PlaybackEventMessageExtension on PlaybackEventMessage {
         duration: duration ?? this.duration,
         icyMetadata: icyMetadata ?? this.icyMetadata,
         currentIndex: currentIndex ?? this.currentIndex,
-        androidAudioSessionId: androidAudioSessionId ?? this.androidAudioSessionId,
+        androidAudioSessionId:
+            androidAudioSessionId ?? this.androidAudioSessionId,
       );
 }
 
@@ -803,7 +859,9 @@ extension AudioSourceExtension on AudioSourceMessage {
     final self = this;
     if (self is ConcatenatingAudioSourceMessage) {
       if (self.id == id) return self;
-      return self.children.map((child) => child.findCat(id)).firstWhere((cat) => cat != null, orElse: () => null);
+      return self.children
+          .map((child) => child.findCat(id))
+          .firstWhere((cat) => cat != null, orElse: () => null);
     } else if (self is LoopingAudioSourceMessage) {
       return self.child.findCat(id);
     } else {
@@ -816,7 +874,9 @@ extension AudioSourceExtension on AudioSourceMessage {
     if (self is ConcatenatingAudioSourceMessage) {
       return self.children.expand((child) => child.sequence).toList();
     } else if (self is LoopingAudioSourceMessage) {
-      return List.generate(self.count, (i) => self.child.sequence).expand((sequence) => sequence).toList();
+      return List.generate(self.count, (i) => self.child.sequence)
+          .expand((sequence) => sequence)
+          .toList();
     } else {
       return [self as IndexedAudioSourceMessage];
     }
@@ -828,7 +888,8 @@ extension AudioSourceExtension on AudioSourceMessage {
       var offset = 0;
       final childIndicesList = <List<int>>[];
       for (final child in self.children) {
-        final childIndices = child.shuffleIndices.map((i) => i + offset).toList();
+        final childIndices =
+            child.shuffleIndices.map((i) => i + offset).toList();
         childIndicesList.add(childIndices);
         offset += childIndices.length;
       }
@@ -855,7 +916,8 @@ class TrackInfo {
   const TrackInfo(this.index, this.duration);
 
   @override
-  bool operator ==(Object other) => other is TrackInfo && index == other.index && duration == other.duration;
+  bool operator ==(Object other) =>
+      other is TrackInfo && index == other.index && duration == other.duration;
 
   @override
   int get hashCode => Object.hash(index, duration);
@@ -868,4 +930,22 @@ class TrackInfo {
 extension _ValueStreamExtension<T> on ValueStream<T> {
   /// Backwards compatible version of valueOrNull.
   T? get nvalue => hasValue ? value : null;
+}
+
+class _PlayingEvent {
+  final bool playing;
+
+  const _PlayingEvent(this.playing);
+}
+
+class _ValueCompleter<T> {
+  final _completer = Completer<T>();
+  T? value;
+
+  void complete(T value) {
+    this.value = value;
+    _completer.complete(value);
+  }
+
+  Future<T> get future => _completer.future;
 }